requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
component==0.0.1
discord.py==2.0.0a3462+g1c40d43f
idna==3.2
multidict==5.1.0
python-dotenv==0.19.0
typing-extensions==3.10.0.0
yarl==1.6.3
